Change how the init of SerialDispatchQueueScheduler and cacheQueue is done.
The init of SerialDispatchQueueScheduler that takes a queue as a parameter, does not use it to do async operations, instead, it creates a new queue that targets the given one. Despite that the Observable ends up running on the specified queue if you want that queue to be used imperatively it will not ensure the enqueue time of an operation.
So by using the SerialDispatchQueueScheduler, that give us through a config closure the reference of the queue, we can ensure that all disk operations are running directly on the same queue without any proxy.