Support TTL to DiskCache.
This is a draft application in DiskCache. I was forced under the current ttl api convention to make the actual get and set of Data to AnyObject and their respective methods internal, also to implement a different set of get/set functions to support the CacheExpirableDTO. Another drawback is that get/set function that returning the CacheExpirableDTO are public due to generic constraints on the ExpirableCache protocol extension to the Cache Protocol.