Introduce async testing of instance retaining in cache composition.
Summary: Tests until now test the tests in a serial manner which makes the test easier to write and they focus mostly in logic.
But if we want tests to resemble more real life scenarios we want to test what is happening when the results are calculates in different queues/threads.

To facilitate the async testing we created the AsyncCacheFake which you can set the delay values when setting and getting results from it.